pyproject.toml
[project]
name = "paw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["python-telegram-bot==21.10"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = ["pytest==8.3.4", "pytest-asyncio==0.25.3"]
